+++ docs/PHASE_PLAN.md
@@ -159,6 +159,9 @@ checked-in docs, tests, and commits only.
 - Typechecker generic behavior-bound validation now lives in a dedicated helper
   module, keeping bound declaration checks and substitution-based impl checks
   out of the broad generic type-reference traversal.
+- Typechecker generic type-reference validation now delegates recursive type,
+  expression, and statement walking to a dedicated helper module, keeping
+  declaration task replay separate from AST traversal details.
 - Resolver validation replay now collects expected declaration symbols,
   expected local symbols, import-validation state, and behavior-association
   replay tasks in one declaration pass before checking resolver-owned extras,
