+++ src/typechecker/generic_type_validation.rs
@@ -462,416 +462,4 @@ impl TypeChecker {
         }
         self.validate_generic_type_ref_bounds(&info.return_type, scoped, span);
     }
-
-    fn validate_generic_type_ref_bounds(
-        &mut self,
-        ast_type: &AstType,
-        scoped_type_params: &HashSet<String>,
-        span: Span,
-    ) {
-        self.validate_generic_type_ref_bounds_with_unknowns(
-            ast_type,
-            scoped_type_params,
-            span,
-            true,
-        );
-    }
-
-    pub(super) fn validate_generic_type_arg_refs_allow_unknowns(
-        &mut self,
-        type_args: &[AstType],
-        span: Span,
-    ) {
-        let scoped_type_params = HashSet::new();
-        self.validate_generic_type_arg_refs_with_unknowns(
-            type_args,
-            &scoped_type_params,
-            span,
-            false,
-        );
-    }
-
-    fn validate_generic_type_arg_refs(
-        &mut self,
-        type_args: &[AstType],
-        scoped_type_params: &HashSet<String>,
-        span: Span,
-    ) {
-        self.validate_generic_type_arg_refs_with_unknowns(
-            type_args,
-            scoped_type_params,
-            span,
-            true,
-        );
-    }
-
-    fn validate_generic_type_arg_refs_with_unknowns(
-        &mut self,
-        type_args: &[AstType],
-        scoped_type_params: &HashSet<String>,
-        span: Span,
-        reject_unknown: bool,
-    ) {
-        for type_arg in type_args {
-            self.validate_generic_type_ref_bounds_with_unknowns(
-                type_arg,
-                scoped_type_params,
-                span,
-                reject_unknown,
-            );
-        }
-    }
-
-    fn validate_generic_type_ref_bounds_with_unknowns(
-        &mut self,
-        ast_type: &AstType,
-        scoped_type_params: &HashSet<String>,
-        span: Span,
-        reject_unknown: bool,
-    ) {
-        match ast_type {
-            AstType::Named(name) => {
-                if scoped_type_params.contains(name) {
-                    return;
-                }
-
-                if !self.is_known_named_type(name) {
-                    if reject_unknown {
-                        self.diagnostics.push(Diagnostic::error(
-                            "E0201",
-                            format!("unknown type symbol '{name}'"),
-                            span,
-                        ));
-                    }
-                    return;
-                }
-
-                let generic = self
-                    .structs
-                    .get(name)
-                    .map(|info| ("struct", info.type_params.len()))
-                    .or_else(|| {
-                        self.enums
-                            .get(name)
-                            .map(|info| ("enum", info.type_params.len()))
-                    });
-                if let Some((kind, type_param_count)) = generic {
-                    if type_param_count > 0 {
-                        self.diagnostics.push(Diagnostic::error(
-                            "E5001",
-                            format!(
-                                "generic {} `{}` expects {} type arguments, found 0",
-                                kind, name, type_param_count
-                            ),
-                            span,
-                        ));
-                    }
-                }
-            }
-            AstType::Generic { name, type_args } => {
-                self.validate_generic_type_arg_refs_with_unknowns(
-                    type_args,
-                    scoped_type_params,
-                    span,
-                    reject_unknown,
-                );
-
-                if scoped_type_params.contains(name) {
-                    return;
-                }
-
-                let (kind, type_params, type_param_bounds) =
-                    if let Some(info) = self.structs.get(name) {
-                        (
-                            "struct",
-                            info.type_params.clone(),
-                            info.type_param_bounds.clone(),
-                        )
-                    } else if let Some(info) = self.enums.get(name) {
-                        (
-                            "enum",
-                            info.type_params.clone(),
-                            info.type_param_bounds.clone(),
-                        )
-                    } else {
-                        if reject_unknown && !self.imports.contains_key(name) {
-                            self.diagnostics.push(Diagnostic::error(
-                                "E0201",
-                                format!("unknown type symbol '{name}'"),
-                                span,
-                            ));
-                        }
-                        return;
-                    };
-
-                if type_params.len() != type_args.len() {
-                    self.diagnostics.push(Diagnostic::error(
-                        "E5001",
-                        format!(
-                            "generic {} `{}` expects {} type arguments, found {}",
-                            kind,
-                            name,
-                            type_params.len(),
-                            type_args.len()
-                        ),
-                        span,
-                    ));
-                    return;
-                }
-
-                let substitutions: HashMap<String, Type> = type_params
-                    .iter()
-                    .zip(type_args.iter())
-                    .filter_map(|(param, arg)| {
-                        if ast_type_references_type_param(arg, scoped_type_params) {
-                            None
-                        } else {
-                            Some((param.clone(), self.resolve_type(arg)))
-                        }
-                    })
-                    .collect();
-                self.check_generic_bounds(&type_param_bounds, &substitutions, span);
-            }
-            AstType::Ptr(inner)
-            | AstType::MutPtr(inner)
-            | AstType::RawPtr(inner)
-            | AstType::Slice(inner) => {
-                self.validate_generic_type_ref_bounds_with_unknowns(
-                    inner,
-                    scoped_type_params,
-                    span,
-                    reject_unknown,
-                );
-            }
-            AstType::Array { elem, .. } => {
-                self.validate_generic_type_ref_bounds_with_unknowns(
-                    elem,
-                    scoped_type_params,
-                    span,
-                    reject_unknown,
-                );
-            }
-            AstType::Function { params, ret } => {
-                self.validate_generic_type_arg_refs_with_unknowns(
-                    params,
-                    scoped_type_params,
-                    span,
-                    reject_unknown,
-                );
-                self.validate_generic_type_ref_bounds_with_unknowns(
-                    ret,
-                    scoped_type_params,
-                    span,
-                    reject_unknown,
-                );
-            }
-            _ => {}
-        }
-    }
-
-    fn is_known_named_type(&self, name: &str) -> bool {
-        self.structs.contains_key(name)
-            || self.enums.contains_key(name)
-            || self.imports.contains_key(name)
-    }
-
-    fn validate_generic_expr_type_references(
-        &mut self,
-        expr: &Expression,
-        scoped_type_params: &HashSet<String>,
-    ) {
-        match expr {
-            Expression::FunctionCall {
-                type_args,
-                args,
-                span,
-                ..
-            } => {
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                for arg in args {
-                    self.validate_generic_expr_type_references(arg, scoped_type_params);
-                }
-            }
-            Expression::MethodCall {
-                receiver,
-                type_args,
-                args,
-                span,
-                ..
-            } => {
-                self.validate_generic_expr_type_references(receiver, scoped_type_params);
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                for arg in args {
-                    self.validate_generic_expr_type_references(arg, scoped_type_params);
-                }
-            }
-            Expression::BinaryOp { left, right, .. } => {
-                self.validate_generic_expr_type_references(left, scoped_type_params);
-                self.validate_generic_expr_type_references(right, scoped_type_params);
-            }
-            Expression::UnaryOp { operand, .. } => {
-                self.validate_generic_expr_type_references(operand, scoped_type_params);
-            }
-            Expression::MemberAccess { object, .. } => {
-                self.validate_generic_expr_type_references(object, scoped_type_params);
-            }
-            Expression::IndexAccess { object, index, .. } => {
-                self.validate_generic_expr_type_references(object, scoped_type_params);
-                self.validate_generic_expr_type_references(index, scoped_type_params);
-            }
-            Expression::StructLiteral {
-                type_args,
-                fields,
-                span,
-                ..
-            } => {
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                for (_, value) in fields {
-                    self.validate_generic_expr_type_references(value, scoped_type_params);
-                }
-            }
-            Expression::EnumVariant {
-                type_args,
-                payload,
-                span,
-                ..
-            } => {
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                if let Some(payload) = payload {
-                    self.validate_generic_expr_type_references(payload, scoped_type_params);
-                }
-            }
-            Expression::ArrayLiteral { elements, .. } => {
-                for element in elements {
-                    self.validate_generic_expr_type_references(element, scoped_type_params);
-                }
-            }
-            Expression::Match {
-                scrutinee, arms, ..
-            } => {
-                self.validate_generic_expr_type_references(scrutinee, scoped_type_params);
-                for arm in arms {
-                    if let Some(guard) = &arm.guard {
-                        self.validate_generic_expr_type_references(guard, scoped_type_params);
-                    }
-                    self.validate_generic_expr_type_references(&arm.body, scoped_type_params);
-                }
-            }
-            Expression::WhileLoop {
-                condition, body, ..
-            } => {
-                self.validate_generic_expr_type_references(condition, scoped_type_params);
-                self.validate_generic_expr_type_references(body, scoped_type_params);
-            }
-            Expression::Loop { body, .. } => {
-                self.validate_generic_expr_type_references(body, scoped_type_params);
-            }
-            Expression::If {
-                condition,
-                then_body,
-                else_body,
-                ..
-            } => {
-                self.validate_generic_expr_type_references(condition, scoped_type_params);
-                self.validate_generic_expr_type_references(then_body, scoped_type_params);
-                if let Some(else_body) = else_body {
-                    self.validate_generic_expr_type_references(else_body, scoped_type_params);
-                }
-            }
-            Expression::Block {
-                statements, expr, ..
-            } => {
-                for statement in statements {
-                    self.validate_generic_statement_type_references(statement, scoped_type_params);
-                }
-                if let Some(expr) = expr {
-                    self.validate_generic_expr_type_references(expr, scoped_type_params);
-                }
-            }
-            Expression::Return { value, .. } => {
-                if let Some(value) = value {
-                    self.validate_generic_expr_type_references(value, scoped_type_params);
-                }
-            }
-            Expression::Closure {
-                params,
-                return_type,
-                body,
-                span,
-            } => {
-                for param in params {
-                    self.validate_generic_type_ref_bounds(
-                        &param.ty,
-                        scoped_type_params,
-                        param.span,
-                    );
-                }
-                if let Some(return_type) = return_type {
-                    self.validate_generic_type_ref_bounds(return_type, scoped_type_params, *span);
-                }
-                self.validate_generic_expr_type_references(body, scoped_type_params);
-            }
-            Expression::Cast {
-                expr,
-                target_type,
-                span,
-            } => {
-                self.validate_generic_expr_type_references(expr, scoped_type_params);
-                self.validate_generic_type_ref_bounds(target_type, scoped_type_params, *span);
-            }
-            Expression::StringInterpolation { parts, .. } => {
-                for part in parts {
-                    if let ast::StringPart::Expr(expr) = part {
-                        self.validate_generic_expr_type_references(expr, scoped_type_params);
-                    }
-                }
-            }
-            Expression::Range { start, end, .. } => {
-                self.validate_generic_expr_type_references(start, scoped_type_params);
-                self.validate_generic_expr_type_references(end, scoped_type_params);
-            }
-            Expression::Defer { expr, .. } => {
-                self.validate_generic_expr_type_references(expr, scoped_type_params);
-            }
-            Expression::IntLiteral { .. }
-            | Expression::FloatLiteral { .. }
-            | Expression::StringLiteral { .. }
-            | Expression::BoolLiteral { .. }
-            | Expression::CharLiteral { .. }
-            | Expression::Identifier { .. }
-            | Expression::Break { .. }
-            | Expression::Continue { .. }
-            | Expression::LoopControl { .. }
-            | Expression::Error { .. } => {}
-        }
-    }
-
-    fn validate_generic_statement_type_references(
-        &mut self,
-        statement: &ast::Statement,
-        scoped_type_params: &HashSet<String>,
-    ) {
-        match statement {
-            ast::Statement::VarDecl {
-                ty, value, span, ..
-            } => {
-                if let Some(ty) = ty {
-                    self.validate_generic_type_ref_bounds(ty, scoped_type_params, *span);
-                }
-                self.validate_generic_expr_type_references(value, scoped_type_params);
-            }
-            ast::Statement::Assignment { target, value, .. } => {
-                self.validate_generic_expr_type_references(target, scoped_type_params);
-                self.validate_generic_expr_type_references(value, scoped_type_params);
-            }
-            ast::Statement::Expression { expr, .. } => {
-                self.validate_generic_expr_type_references(expr, scoped_type_params);
-            }
-            ast::Statement::Block { stmts, .. } => {
-                for statement in stmts {
-                    self.validate_generic_statement_type_references(statement, scoped_type_params);
-                }
-            }
-        }
-    }
 }
